# openapi_preview.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Script Rating API",
        "description": "API for rating and analyzing scripts using clean architecture",
        "version": "0.1.0"
    },
    "paths": {
...
        "/api/documents/upload": {
            "post": {
                "tags": [
                    "Documents"
                ],
                "summary": "Upload a document",
                "description": "Upload script or criteria documents. Criteria files are indexed into the knowledge base.",
                "operationId": "upload_document_api_documents_upload_post",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/Body_upload_document_api_documents_upload_post"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DocumentUploadResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/documents/{document_id}": {
            "get": {
                "tags": [
                    "Documents"
                ],
                "summary": "Get document metadata",
                "description": "Retrieve metadata for an uploaded document.",
                "operationId": "get_document_api_documents__document_id__get",
                "parameters": [
                    {
                        "name": "document_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Document Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Documents"
                ],
                "summary": "Delete a document",
                "description": "Delete an uploaded document and its associated data.",
                "operationId": "delete_document_api_documents__document_id__delete",
                "parameters": [
                    {
                        "name": "document_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Document Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Body_upload_document_api_documents_upload_post": {
                "properties": {
                    "file": {
                        "type": "string",
                        "format": "binary",
                        "title": "File"
                    },
                    "filename": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Filename"
                    },
                    "document_type": {
                        "$ref": "#/components/schemas/DocumentType",
                        "default": "script"
                    }
                },
                "type": "object",
                "required": [
                    "file"
                ],
                "title": "Body_upload_document_api_documents_upload_post"
            },
...
            "DocumentType": {
                "type": "string",
                "enum": [
                    "script",
                    "criteria"
                ],
                "title": "DocumentType"
            },
            "DocumentUploadResponse": {
                "properties": {
                    "document_id": {
                        "type": "string",
                        "title": "Document Id",
                        "description": "Unique identifier for the uploaded document"
                    },
                    "filename": {
                        "type": "string",
                        "title": "Filename",
                        "description": "Name of the uploaded file"
                    },
                    "uploaded_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Uploaded At",
                        "description": "Timestamp of upload"
                    },
                    "document_type": {
                        "$ref": "#/components/schemas/DocumentType",
                        "description": "Type of uploaded document (script or criteria)"
                    },
                    "chunks_indexed": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Chunks Indexed",
                        "description": "Number of knowledge chunks indexed for criteria documents"
                    },
                    "status": {
                        "type": "string",
                        "title": "Status",
                        "description": "Upload status",
                        "default": "uploaded"
                    }
                },
                "type": "object",
                "required": [
                    "document_id",
                    "filename",
                    "uploaded_at",
                    "document_type"
                ],
                "title": "DocumentUploadResponse",
                "description": "Response model for successful document upload."
            },
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            },
...
        }
    }
}